temporary fix is applied for test purpose

+++ app/src/main/java/foundation/e/apps/manager/database/fusedDownload/FusedDownloadDAO.kt
@@ -27,7 +27,7 @@ interface FusedDownloadDAO {
     fun getDownloadFlowById(id: String): LiveData<FusedDownload>
 
     @Update
-    suspend fun updateDownload(fusedDownload: FusedDownload): Long
+    suspend fun updateDownload(fusedDownload: FusedDownload): Int
 
     @Delete
     suspend fun deleteDownload(fusedDownload: FusedDownload)
